backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
filelock==3.18.0
flake8==7.2.0
gunicorn==23.0.0
identify==2.6.12
inflection==0.5.1
isort==6.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.31.2
